+++ src/main/java/com/ccoupang/orderproduct/OrderProduct.java
@@ -1,8 +1,7 @@
-package com.ccoupang.domain.review;
+package com.ccoupang.orderproduct;
 
-import com.ccoupang.domain.member.Member;
-import com.ccoupang.domain.product.Product;
-import java.time.LocalDateTime;
+import com.ccoupang.order.Order;
+import com.ccoupang.product.Product;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -14,27 +13,25 @@ import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
 @Entity
-@Table(name = "review")
-public class Review {
+@Table(name = "order_product")
+public class OrderProduct {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "review_id")
+  @Column(name = "order_product_id")
   private Long id;
 
   @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "member_id", nullable = false)
-  private Member member;
+  @JoinColumn(name = "order_id", nullable = false)
+  private Order order;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "product_id", nullable = false)
   private Product product;
 
   @Column(nullable = false)
-  private LocalDateTime reviewDate;
+  private int quantity;
 
   @Column(nullable = false)
-  private String description;
-
-  private int rate_scoer;
+  private int price;
 }
