+++ src/main/java/com/ccoupang/payment/PaymentRepository.java
@@ -1,4 +1,4 @@
-package com.ccoupang.domain.payment;
+package com.ccoupang.payment;
 
 import org.springframework.data.jpa.repository.JpaRepository;
