+++ src/main/java/com/ccoupang/payment/PaymentStatus.java
@@ -1,4 +1,4 @@
-package com.ccoupang.domain.payment;
+package com.ccoupang.payment;
 
 public enum PaymentStatus {
   CARD,
